fix: Swift FFI 类型不匹配 Int vs UInt 导致 iOS 单测构建失败
FFI 函数的 uintptr_t 参数在 Swift 中导入为 UInt，但代码传入 Int。
- Sha256.swift: rawBuffer.count → UInt(rawBuffer.count)
- Token.swift: aether_estimate_tokens 返回 UInt，需 Int(...) 转换
- Vector.swift: aBuf.count/bBuf.count → UInt(...)（cosine_f32/f64 共 4 处）
- Chunker.swift: maxChars/overlapChars → UInt(...)（同问题）

+++ Packages/AetherCore/Sources/AetherRust/Chunker.swift
@@ -13,7 +13,7 @@ public enum AetherRustChunker {
     /// 与 `DocumentChunker.chunkDocument` 的纯 Swift 实现行为一致。
     public static func chunkDocument(_ text: String, maxChars: Int, overlapChars: Int) -> [String] {
         return text.withCString { ptr in
-            guard let raw = aether_chunk_document(ptr, maxChars, overlapChars) else {
+            guard let raw = aether_chunk_document(ptr, UInt(maxChars), UInt(overlapChars)) else {
                 return [text]
             }
             defer { aether_free_string(raw) }
